requirements.txt
annotated-types==0.6.0
anyio==3.7.1
autopep8==2.0.4
blinker==1.7.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==41.0.7
ecdsa==0.18.0
fastapi==0.105.0
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
greenlet==3.0.2
gunicorn==21.2.0
h11==0.14.0
httptools==0.6.1
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.2
pyasn1==0.5.1
pycodestyle==2.11.1
pycparser==2.21
pydantic==2.5.2
pydantic-settings==2.1.0
pydantic_core==2.14.5
pyodbc==5.0.1
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
starlette==0.27.0
typing_extensions==4.9.0
urllib3==2.1.0
uvloop==0.19.0
uvicorn==0.24.0.post1
Werkzeug==3.0.1
